64 bit build
git-svn-id: svn://svn.code.sf.net/p/ditto-cp/code/trunk@561 595ec19a-5cb4-439b-94a8-42fb3063c22c

+++ ClipFormatQListCtrl.h
@@ -8,7 +8,7 @@ public:
 	CClipFormatQListCtrl(void);
 	~CClipFormatQListCtrl(void);
 
-	long m_clipRow;
+	int m_clipRow;
 	bool m_convertedToSmallImage;
 
 	HGLOBAL GetDib(CDC *pDc, int height);
